refactor: input component 이름 수정에 따른 수정 및 props 수정
